requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
accelerate=0.34.2=pypi_0
aiohappyeyeballs=2.5.0=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
anyio=4.6.2=py311haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h827c3e9_1
arrow=1.3.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py311haa95532_0
attrs=25.1.0=pypi_0
babel=2.16.0=py311haa95532_0
beautifulsoup4=4.12.3=py311haa95532_0
bitsandbytes=0.43.1=pypi_0
blas=1.0=mkl
bleach=6.2.0=py311haa95532_0
brotli-python=1.0.9=py311h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=py311haa95532_0
cffi=1.17.1=py311h827c3e9_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.8=pypi_0
colorama=0.4.6=py311haa95532_0
comm=0.2.1=py311haa95532_0
cuda-cccl=12.8.90=0
cuda-cccl_win-64=12.8.90=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.8.90=0
cuda-runtime=11.8.0=0
cuda-version=12.8=3
datasets=2.20.0=pypi_0
debugpy=1.8.11=py311h5da7b33_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.8=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.1=pypi_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py311haa95532_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.5.0=pypi_0
fsspec=2024.5.0=pypi_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py311h827c3e9_0
google-auth=2.38.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.70.0=pypi_0
h11=0.14.0=py311haa95532_0
httpcore=1.0.2=py311haa95532_0
httpx=0.27.0=py311haa95532_0
huggingface-hub=0.29.2=pypi_0
icu=73.1=h6c2663c_0
idna=3.7=py311haa95532_0
intel-openmp=2021.4.0=pypi_0
ipykernel=6.29.5=py311haa95532_1
ipython=8.30.0=py311haa95532_0
ipywidgets=8.1.5=py311haa95532_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=py311haa95532_0
jieba=0.42.1=pypi_0
jinja2=3.1.5=py311haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
json5=0.9.25=py311haa95532_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=py311haa95532_0
jsonschema-specifications=2023.7.1=py311haa95532_0
jupyter=1.1.1=py311haa95532_0
jupyter-lsp=2.2.0=py311haa95532_0
jupyter_client=8.6.3=py311haa95532_0
jupyter_console=6.6.3=py311haa95532_0
jupyter_core=5.7.2=py311haa95532_0
jupyter_events=0.12.0=py311haa95532_0
jupyter_server=2.15.0=py311haa95532_0
jupyter_server_terminals=0.4.4=py311haa95532_1
jupyterlab=4.3.4=py311haa95532_0
jupyterlab_pygments=0.3.0=py311haa95532_0
jupyterlab_server=2.27.3=py311haa95532_0
jupyterlab_widgets=3.0.13=py311haa95532_0
krb5=1.20.1=h5b6d351_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.9.90=0
libcurand-dev=10.3.9.90=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libpq=17.4=h70ee33d_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=h44ae7cf_1
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py311h827c3e9_0
matplotlib-inline=0.1.6=py311haa95532_0
mdurl=0.1.2=pypi_0
mistune=3.1.2=py311haa95532_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py311h2bbff1b_0
mkl_fft=1.3.1=py311h743a336_0
mkl_random=1.2.2=py311heda8569_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py311haa95532_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
nbclient=0.10.2=py311haa95532_0
nbconvert=7.16.6=py311haa95532_0
nbconvert-core=7.16.6=py311haa95532_0
nbconvert-pandoc=7.16.6=py311haa95532_0
nbformat=5.10.4=py311haa95532_0
nest-asyncio=1.6.0=py311haa95532_0
networkx=3.4.2=py311haa95532_0
nltk=3.9.1=pypi_0
notebook=7.3.2=py311haa95532_0
notebook-shim=0.2.3=py311haa95532_0
numpy=1.24.3=py311ha6a8073_0
numpy-base=1.24.3=py311h877bb12_0
oauthlib=3.2.2=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.16=h3f729d1_0
overrides=7.4.0=py311haa95532_0
packaging=24.2=py311haa95532_0
pandas=2.2.3=pypi_0
pandoc=2.12=haa95532_3
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.4=py311haa95532_0
peft=0.15.0=pypi_0
pillow=11.1.0=py311h096bfcc_0
pip=25.0=py311haa95532_0
platformdirs=3.10.0=py311haa95532_0
prometheus_client=0.21.0=py311haa95532_0
prompt-toolkit=3.0.43=py311haa95532_0
prompt_toolkit=3.0.43=hd3eb1b0_0
propcache=0.3.0=pypi_0
protobuf=6.30.0=pypi_0
psutil=7.0.0=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=19.0.1=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.19.1=pypi_0
pyqt=6.7.1=py311h5da7b33_0
pyqt6-sip=13.9.1=py311h827c3e9_0
pysocks=1.7.1=py311haa95532_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0post0=py311haa95532_2
python-fastjsonschema=2.20.0=py311haa95532_0
python-json-logger=3.2.1=py311haa95532_0
pytorch=2.3.0=py3.11_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2025.1=pypi_0
pywin32=308=py311h5da7b33_0
pywinpty=2.0.14=py311h72d21ff_0
pyyaml=6.0.2=py311h827c3e9_0
pyzmq=26.2.0=py311h5da7b33_0
qtbase=6.7.2=h0804d20_1
qtconsole=5.6.1=py311haa95532_0
qtdeclarative=6.7.2=h5da7b33_0
qtpy=2.4.1=py311haa95532_0
qtsvg=6.7.2=hf2fb9eb_0
qttools=6.7.2=h0de5f00_0
qtwebchannel=6.7.2=h5da7b33_0
qtwebsockets=6.7.2=h5da7b33_0
referencing=0.30.2=py311haa95532_0
regex=2024.11.6=pypi_0
requests=2.32.3=py311haa95532_1
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=py311haa95532_0
rfc3986-validator=0.1.1=py311haa95532_0
rich=13.9.4=pypi_0
rouge-chinese=1.0.3=pypi_0
rpds-py=0.22.3=py311h636fa0f_0
rsa=4.9=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
send2trash=1.8.2=py311haa95532_1
sentencepiece=0.2.0=pypi_0
setuptools=75.8.0=py311haa95532_0
shtab=1.7.1=pypi_0
sip=6.10.0=py311h5da7b33_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post12=pypi_0
sniffio=1.3.0=py311haa95532_0
soupsieve=2.5=py311haa95532_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.3=py311haa95532_1
tbb=2021.13.1=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.17.1=py311haa95532_0
threadpoolctl=3.6.0=pypi_0
tinycss2=1.4.0=py311haa95532_0
tk=8.6.14=h0416ee5_0
tokenizers=0.21.1=pypi_0
torchaudio=2.3.0=pypi_0
torchvision=0.18.0=pypi_0
tornado=6.4.2=py311h827c3e9_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=py311haa95532_0
transformers=4.50.0=pypi_0
trl=0.8.6=pypi_0
typeguard=4.4.2=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.12.2=py311haa95532_0
typing_extensions=4.12.2=py311haa95532_0
tyro=0.9.16=pypi_0
tzdata=2025.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.3.0=py311haa95532_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=py311haa95532_1
websocket-client=1.8.0=py311haa95532_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py311haa95532_0
widgetsnbextension=4.0.13=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
winpty=0.4.3=4
xxhash=3.5.0=pypi_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=hd77b12b_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
